fixed migrations

+++ db/migrate/20140408084305_remove_info_requests.rb
@@ -1,6 +1,6 @@
-class CreateInfoRequests < ActiveRecord::Migration
+class RemoveInfoRequests < ActiveRecord::Migration
   def up
-    drop_table :info_requests
+    drop_table :info_requests if ActiveRecord::Base.connection.table_exists? :info_requests
   end
 
   def down
